+++ app/src/main/java/com/hirauchi/todo_matrix/model/ToDo.kt
@@ -3,6 +3,6 @@ package com.hirauchi.todo_matrix.model
 data class ToDo(
     val id: Int,
     val content: String,
-    val importance: Float,
-    val urgency: Float
+    val importance: Int,
+    val urgency: Int
 )
